settings: Add new no_strict_options setting
Added a new `no_strict_options` `Setting`. This changes the way the
parser works subtly: the default works like `getopt(3)` such that if an
option is specified as taking an argument, the next command-line
argument will become that argument (even if it starts with a dash).
Setting `no_strict_options=true` disables this behaviour such that if
the argument after an option requiring an argument itself starts with a
dash, it will be considered an option.

Signed-off-by: James O. D. Hunt <jamesodhunt@gmail.com>